+++ WeatherApi_CI-CD/Properties/launchSettings.json
@@ -14,7 +14,8 @@
       "dotnetRunMessages": true,
       "launchBrowser": true,
       "launchUrl": "swagger",
-      "applicationUrl": "https://dev.kjeld.io:40200;http://dev.kjeld.io:20200",
+      //"applicationUrl": "https://dev.kjeld.io:40201;http://dev.kjeld.io:20200",
+      "applicationUrl": "http://localhost:5290",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
